Initial support for linchpin integration
* Made 'os' import statements consistent between entry_point.py and
wrappers.py

* Fixed some line length issues in cinch() in entry_point.py

* Added 'cinchpin' command and associated CLI arguments.  This command can
be called the same way as 'linchpin', e.g. "cinchpin rise" or optionally
you can provide a path to the linch-pin working directory, e.g.
"cinchpin rise -w /path/to/workdir".

* Added "pushd" work-around using a contextmanager to allow linch-pin to
be in the correct working directory to operate

* Fixed some line length issues in call_ansible()

* Added StrictHostKeyChecking=no as a default setting to aid in
automation with call_ansible().  This removes the need for ansible.cfg
entirely.

* Created a generic command_handler() function so that call_ansible()
and call_linchpin() can share some code.

* Added call_linchpin() to find necessary data within in the linch-pin
working directory so that call_ansible can be passed the necessary
inventory file when it is called after provisioning.  This function
supports a subset of linchpin commands (currently only the 'rise'
subcommand is available, but more will be added in the future).

* Added an Ansible role named "check_ssh" that is a work-around to
handle waiting on cloud-init to create a working SSH configuration on
the server side.  Without this, Ansible will fail pretty spectactularly
and surprisingly doesn't have a built-in way to simply retry SSH
connections until authentication is successful.