linchpin 1.0.x compatibility
* Removed 'pushd/pop' work-around from wrappers.py now that linchpin has
the -w option on the CLI

* Updated docs links in wrappers.py to make the actual content easier to
find

* We don't need to raise an exception for unsupported linchpin commands
in cinchpin.  It's better to just print a normal error message for the
user.

* cinchpin now supports the new CLI options of linchpin and removes the
old ones

* Added some default linchpin CLI args that should be useful, such as
printing Ansible output, pointing to the correct workspace directory,
and setting the creds path to that of the workspace directory

* Added group_vars support to 'cinchpin init'

* Updated local_paths dictionary to account for the new group_vars
directory and to create the proper skeleton info

* Removed redundant places where we print an error and sys.exit(1).  We
can use sys.exit('msg') and get the same result.

* Vendored jenkins_script Ansible module because linchpin v1.0.x doesn't
depend on a recent enough version of Ansible to include the module
within Ansible itself

* Added the ability to optionally install the latest tip of
cinch+linchpin via the install-rhel7 Ansible playbook

* Added beaker-client package to install-rhel7 Ansible playbook

* Removed a couple of old work-arounds that from install-rhel7 Ansible
playbook that are no longer needed as of linchpin 1.0.x

* Added check_ssh role to teardown.yml.  This was missing before and was
a bug that was exposed under some conditions.

* Updated users.rst documention to reflect all the changes with linchpin
1.0.x and cinchpin

* JJB template now checks out example topologies into a subdirectory of
the Jenkins $WORKSPACE, which should be best practice

* Added $PROVIDER parameter to example JJB template, which allows for
easy switching between OpenStack and Beaker examples

* Removed work-arounds from JJB template which no longer apply due to
linchpin 1.0.x

* Updated JJB template to use the new cinchpin commands

* Added linchpin 'outputs' files as artifacts for the JJB template,
which are generated during provisioning and must be passed to the
teardown jobs

* The 'runtest' portion of the JJB template has a more meaningful
example other than just 'exit 0'

* Added outputs files to the copyartifact portion of the JJB template so
that teardown jobs can run as expected.  There's also a symlink
work-around for when some linchpin providers name their outputs file
with a different file extension.

* Bumped version in setup.py to 0.7.0, with a requirement for linchpin
1.0.1

* Updated path in JJB installer for cinch to point to the new
cinch/playbooks directory

* Removed extraneous {{ python }} var from the task in the
install-rhel7.yml playbook that grabs the latest tip of cinch+linchpin