Remove id_rsa file
Having a private SSH key file on a public repository was a bad idea
from the start. However, one is needed for running Jenkins stuff. This
change will generate a new one, when one is not already present in the
{{ jenkins_user_home }}/.ssh/id_rsa location, but it will not force
you to provide a file beforehand, nor will it overwrite your file if it
already exists in that locale.