Clean up python-pip
Python pip is common to both Jenkins masters and slaves. Making it
common to both should be straightforward here. However, it is named
python2-pip in both EPEL7 and Fedora >= 26. But in Fedora 25, which is
technically still in support it is named python-pip, instead. So, we
need to support that.