Lighten up on facts
Reduces the burden of gather_facts on simple file uploads

Caveat: this means using {{ ansible_user_dir }} and similar facts in
variables passed into this role are going to present additional
challenges

Removed pre-defined values for pre/post-_upload_files out of the
group_vars/all file and into filter defaults. This is because of the
precedence order for variables in group_vars/all, which made this tough
to test with Vagrant.

Addresses #218